// app/Fantasy/frontend/src/models/Player.model.ts
export type Player = {
  id: number;
  firstName: string;
  lastName: string;
  country: string;
  team: string;
  age: number;
  position: string;
  height: string;
  weight: string;
  photoUrl: string;
  gamesPlayed: number;
  pointsPerGame: number;
  reboundsPerGame: number;
  assistsPerGame: number;
  stealsPerGame: number;
  blocksPerGame: number;
  fieldGoalPercentage: number;
  fieldGoalMade: number;
  fieldGoalMiss: number;
  threePointPercentage: number;
  threePointMade: number;
  threePointMiss: number;
  twoPointsPercentage: number;
  twoPointMade: number;
  twoPointMiss: number;
  freeThrowPercentage: number;
  freeThrowMade: number;
  freeThrowMiss: number;
  plusMinusIndex: number;
  pir: number;
  turnoversPerGame: number;
  minutesPerGame: number;
  jerseyNumber: number;
};